requirements.txt
python-dotenv>=1.0.1
numpy>=1.26.0
scikit-learn>=1.4.0
pandas>=2.2.0
SQLAlchemy>=2.0.0
pymysql>=1.1.0
mysqlclient>=2.2.1
openai>=1.30.0
django>=4.2,<5
pypdf>=4.0.0
python-docx>=1.1.0
langchain
langgraph
langchain-openai
langchain-community
langchain-chroma
chromadb
faiss-cpu
langchain-huggingface
langchain-ollama
duckduckgo-search
rank-bm25
